+++ packages/jugg/CHANGELOG.md
@@ -1,9 +1,8 @@
-# CHANGELOG
-
 ## 0.2.1-alpha.2
 
 - feat: 插件可以扩展用户配置
 - refactor: `quiet` of `webpack-dev-server` always true
+- refactor: 配置文件统一管理，如 `tsconfig.json`
 
 ## 0.2.1-alpha.1 - 2020-01-02